RHINENG-5152 refine entitiesReducer.js file
